+++ apina-core/src/main/kotlin/fi/evident/apina/model/settings/EnumMode.kt
@@ -1,6 +1,7 @@
 package fi.evident.apina.model.settings
 
 enum class EnumMode {
-    ENUM,
-    STRING
+    DEFAULT,
+    INT_ENUM,
+    STRING_UNION
 }